+++ ai-test/src/prompts.rs
@@ -1,9 +1,8 @@
-use std::{path::Path, fs};
+use std::{fs, path::Path};
 
 use anyhow::Error;
 
-const PROMPTS_YAML: & str = include_str!("../data/prompts.yaml");
-
+const PROMPTS_YAML: &str = include_str!("../data/prompts.yaml");
 
 pub fn read(path: Option<impl AsRef<Path>>) -> Result<Vec<Vec<String>>, Error> {
     if let Some(path) = path {
@@ -11,5 +10,4 @@ pub fn read(path: Option<impl AsRef<Path>>) -> Result<Vec<Vec<String>>, Error> {
     } else {
         Ok(serde_yaml::from_str(PROMPTS_YAML)?)
     }
-
 }
