+++ src/ai/mod.rs
@@ -1,6 +1,6 @@
-pub mod help;
 pub mod constants;
 pub mod embeddings;
 pub mod error;
 pub mod explain;
+pub mod help;
 pub mod helpers;
